Fix build failure in the repo
* Use 1.3 Zlib version
- The built-in base line does not work, since it will try to get zlib from https://www.zlib.net/zlib-1.2.12.tar.gz,
which is no longer valid. To fix this, I changed ZLIB version to 1.3.

- able to attempt git clone with `ARROW_GIT_REPOSITORY1` being used instead of `ARROW_GIT_REPOSITORY`. The root cause is unknown, need to investigate.

Fix ARROW_GIT_REPOSITORY not set error

In `flight_sql\CMakeLists.txt`, set(ARROW_GIT_REPOSITORY ... CACHE STRING) only sets the default value of ARROW_GIT_REPOSITORY. This value is supposed to be set in the build script (e.g.,  `build_win64.bat`), and the setting of this value is fixed.